backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-generativeai>=0.8.0
google-cloud-texttospeech>=2.16.0
python-dotenv>=1.0.0
gtts>=2.5.0
sounddevice>=0.4.6
numpy>=1.26.0
scipy>=1.13.0
soundfile>=0.12.1
httpx>=0.27.0
